+++ packages/playback/src/lib/consts/interceptor-type.ts
@@ -2,4 +2,5 @@
 // since they can be used as values they should not be in the types folder
 export enum InterceptorType {
   NetworkRequest = 'NetworkRequest',
+  HlsPlaylistParse = 'HlsPlaylistParse',
 }
