--- packages/playback/src/lib/types/interceptor-type-to-interceptor-map.declarations.ts
@@ -1,5 +1,0 @@
-import type { InterceptorType } from '../consts/interceptor-type';
-
-export interface InterceptorTypeToInterceptorMap {
-  [InterceptorType.NetworkRequest]: (request: Request) => Promise<Request>;
-}
