+++ packages/playback/src/lib/types/mappers/interceptor-type-to-interceptor-map.declarations.ts
@@ -1,0 +1,6 @@
+import type { InterceptorType } from '../../consts/interceptor-type';
+
+export interface InterceptorTypeToInterceptorMap {
+  [InterceptorType.NetworkRequest]: (request: Request) => Promise<Request>;
+  [InterceptorType.HlsPlaylistParse]: (playlist: Uint8Array) => Promise<Uint8Array>;
+}
